+++ internal/tests/context.go
@@ -1,0 +1,8 @@
+package tests
+
+import (
+	"context"
+	"reflect"
+)
+
+var ContextType = reflect.TypeOf((*context.Context)(nil)).Elem()
